+++ emails/_components/Params.ts
@@ -20,7 +20,7 @@ export const samplePublicRequest: PublicCloudRequestWithRequestedProject = {
   decisionMakerEmail: 'manager@example.com',
   type: 'CREATE', // or EDIT
   decisionStatus: 'PENDING', // or APPROVED, REJECTED, PROVISIONED
-  humanComment: 'This is a sample request',
+  adminComment: 'This is a sample request',
   active: true,
   created: new Date('2023-11-30T00:00:00Z'),
   decisionDate: new Date('2023-12-01T00:00:00Z'),
@@ -83,7 +83,7 @@ export const samplePublicEditRequest: PublicCloudRequestWithProjectAndRequestedP
   decisionMakerEmail: 'decider@example.com',
   type: 'CREATE', // or EDIT
   decisionStatus: 'PENDING', // or APPROVED, REJECTED, PROVISIONED
-  humanComment: 'This is an example request',
+  adminComment: 'This is an example request',
   active: true,
   created: new Date('2023-11-30T00:00:00Z'),
   decisionDate: new Date('2023-12-01T00:00:00Z'),
@@ -259,7 +259,7 @@ export const samplePrivateRequest: PrivateCloudRequestWithRequestedProject = {
   decisionDate: sampleDate,
   isQuotaChanged: false,
   projectId: null,
-  userComment: 'Some comment',
+  userComment: 'Some comment regarding a request by the user',
   requestedProjectId: 'f',
   userRequestedProjectId: 'e',
   requestedProject: {
@@ -349,7 +349,7 @@ export const samplePrivateEditRequest: PrivateCloudRequestWithProjectAndRequeste
   created: sampleDate,
   isQuotaChanged: false,
   decisionDate: sampleDate,
-  userComment: 'Some comment',
+  userComment: 'Some comment made by the user',
   projectId: null,
   requestedProjectId: 'f',
   userRequestedProjectId: 'e',